Updated some order features

+++ src/main/java/dev/mrkevr/ecommerce/dto/UserProfileResponse.java
@@ -45,7 +45,7 @@ public class UserProfileResponse {
 	
 	private int totalCartItems;
 	
-	private int totalActiveOrders;
+	private long totalActiveOrders;
 	
 	private LocalDateTime created;
